Use symmetryXyXzYzType for the symmetry attribute in ductType
We noticed a minor inconsistency in the symmetry attribute of the ductType while updating TiGL to CPACS 3.4. 

The ductType was created by us on the basis of CPACS 3.2 (I think) and therefore it does not include the changes provided in https://github.com/DLR-SL/CPACS/issues/735, where the symmetry attributes where changed to the new complexType `symmetryXyXzYzType`.